+++ src/uarm/uarm/pxa_DMA.cpp
@@ -53,7 +53,7 @@ struct SocDma {
     void DoSaveLoad(T& chunkHelper) {
         chunkHelper.Do32(dalgn).Do32(dpcsr).Do32(DINT).DoBuffer(CMR, sizeof(CMR));
 
-        for (uint8_t i = 0; i < 32; i++) channels[i].DoSaveLoad(chunkHelper);
+        for (size_t i = 0; i < 32; i++) channels[i].DoSaveLoad(chunkHelper);
     }
 };
